Restructure session commands under rocky sessions group
Move Start, Stop, Status, Edit into Sessions command group as
subcommands. Keep start, stop, status as hidden top-level shortcuts.
Remove edit as a top-level command.
